requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-extensions==3.2.3
djangorestframework==3.16.0
Flask==3.0.3
Flask-WTF==1.2.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
unicorn==2.0.1.post1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
